refactor: TT-157 데이터 목록 ui 분리
- 테마 데이터 목록을 보여주는 부분을 widgets 폴더로 분리
- 검색 path에 따라 상황에 맞는 데이터 목록 widget을 보여줄 예정

#TT-102 practiceHistory
