+++ lib/features/practice_history/controller/history_controller.dart
@@ -2,7 +2,7 @@ import 'package:get/get.dart';
 import 'package:swm_peech_flutter/features/practice_history/data_source/history_theme_data_source.dart';
 import 'package:swm_peech_flutter/features/practice_history/model/history_theme_model.dart';
 
-class ThemeListCtr extends GetxController {
+class HistoryCtr extends GetxController {
 
   Rx<List<HistoryThemeModel>?> themeList = Rx<List<HistoryThemeModel>?>(null);
