+++ lib/routers/routers.dart
@@ -1,7 +1,7 @@
 import 'package:get/get.dart';
 import 'package:swm_peech_flutter/features/home/view/home_screen.dart';
-import 'package:swm_peech_flutter/features/practice_history/controller/theme_list_controller.dart';
-import 'package:swm_peech_flutter/features/practice_history/view/theme_list_screen.dart';
+import 'package:swm_peech_flutter/features/practice_history/controller/history_controller.dart';
+import 'package:swm_peech_flutter/features/practice_history/view/history_screen.dart';
 import 'package:swm_peech_flutter/features/script_input/controller/script_input_controller.dart';
 import 'package:swm_peech_flutter/features/script_input/view/script_input_screen.dart';
 import 'package:swm_peech_flutter/features/theme_input/controller/theme_input_controller.dart';
@@ -49,9 +49,9 @@ class Routers {
     ),
     GetPage(
       name: '/historyThemeList',
-      page: () => const ThemeListScreen(),
+      page: () => const HistoryScreen(),
       binding: BindingsBuilder(() {
-        Get.lazyPut<ThemeListCtr>(() => ThemeListCtr());
+        Get.lazyPut<HistoryCtr>(() => HistoryCtr());
       })
     )
   ];
